✨  상세 페이지에 평균 별점, 별점 별 개수, 총 별점 개수를 반환하는 api 구현
 reviewRepository에서 조회하는 api들의 url 경로를 smoking-area에서 reviews로 변경했습니다.

+++ src/main/java/com/ssmoker/smoker/domain/review/dto/ReviewStarsInfoResponse.java
@@ -2,5 +2,5 @@ package com.ssmoker.smoker.domain.review.dto;
 
 import java.util.List;
 
-public record ReviewStarsInfoResponse(List<Double> stars, double avg, int count) {
+public record ReviewStarsInfoResponse(List<Integer> stars, double avg, int count) {
 }
